requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.19.1
pillow==10.1.0
numpy==1.26.2